pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "quantdatasource"
version = "2.0.0"
authors = [
    {name = "lzq", email = 'justgotpaid88@qq.com'}
]
description = 'finance datasource downloader and importer'
requires-python = ">=3.9"
dependencies = ["numpy>=1.26.4", "pymongo", "pyyaml", "apscheduler", "pandas", "tushare", "quantcalendar", "quantdata", "tqsdk"]

[project.optional-dependencies]
more = ["taos", "baostock"]

[tool.setuptools.package-dir]
quantdatasource= "quantdatasource"
